Restructure client's src/ folder.
