+++ client/src/components/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Redirect, Route, RouteComponentProps, RouteProps } from 'react-router';
 import { useLogin } from '../hooks/LoginService';
-import { RoutingPath } from '../util/routing/Routing.routes';
+import { RoutingPath } from '../routes/Routing.routes';
 
 interface PrivateRouteProps extends RouteProps {
   // Make the component required in a private route!
