+++ client/src/view/Login.tsx
@@ -9,7 +9,7 @@ import ChangePasswordForm, {
 import LoginForm, { LoginFormState } from '../components/forms/LoginForm';
 import { useLogin } from '../hooks/LoginService';
 import { FormikSubmitCallback } from '../types';
-import { PATH_REDIRECT_AFTER_LOGIN } from '../util/routing/Routing.routes';
+import { PATH_REDIRECT_AFTER_LOGIN } from '../routes/Routing.routes';
 import { Information as InfoIcon } from 'mdi-material-ui';
 
 const useStyles = makeStyles((theme: Theme) =>
