+++ client/src/view/points-sheet/enter-form/EnterPoints.helpers.ts
