+++ client/src/view/points-sheet/enter-form/components/EnterPointsForm.helpers.ts
