+++ client/src/view/points-sheet/enter-form/components/ExerciseBox.tsx
