+++ client/src/view/points-sheet/overview/components/TeamCardList.tsx
