+++ client/src/view/points-sheet/overview/components/TeamCardPointsTable.tsx
