+++ client/src/view/points-sheet/util/helper.ts
